+++ api/package.json
@@ -7,9 +7,9 @@
       "build": "npm run prisma:generate && npx tsc --build --clean && npx tsc -p tsconfig.build.json",
       "dev": "nodemon",
       "start": "npm run build && node dist/index.js",
-    "prisma:migrate:core": "prisma migrate dev --schema ./prisma/core/schema.core.prisma",
-    "prisma:migrate:analytics": "prisma migrate dev --schema ./prisma/analytics/schema.analytics.prisma",
-    "prisma:migrate:deploy": "prisma migrate deploy --schema ./prisma/core/schema.core.prisma && prisma migrate deploy --schema ./prisma/analytics/schema.analytics.prisma",
+      "prisma:migrate:core": "prisma migrate dev --schema ./prisma/core/schema.core.prisma",
+      "prisma:migrate:analytics": "prisma migrate dev --schema ./prisma/analytics/schema.analytics.prisma",
+      "prisma:migrate:deploy": "prisma migrate deploy --schema ./prisma/core/schema.core.prisma && prisma migrate deploy --schema ./prisma/analytics/schema.analytics.prisma",
       "prisma:generate": "prisma generate --schema ./prisma/core/schema.core.prisma && prisma generate --schema ./prisma/analytics/schema.analytics.prisma",
       "prettier": "prettier --write . --plugin=prettier-plugin-organize-imports",
       "lint": "eslint . --ext .ts,.tsx",
