+++ api/src/jobs/export-organizations-to-pg/utils/transformer.ts
@@ -1,0 +1,55 @@
+import { Organization as PgOrganization } from "../../../db/analytics";
+import { Organization as MongoOrganization } from "../../../types";
+import { slugify } from "../../../utils";
+
+export const transformMongoOrganizationToPg = (doc: MongoOrganization) => {
+  const obj = {
+    old_id: doc._id.toString(),
+    rna: doc.rna,
+    siren: doc.siren || doc.siret?.slice(0, 9),
+    siret: doc.siret,
+    rup_mi: doc.rupMi,
+    gestion: doc.gestion,
+    status: doc.status,
+    created_at: doc.createdAt,
+    last_declared_at: doc.lastDeclaredAt,
+    published_at: doc.publishedAt,
+    dissolved_at: doc.dissolvedAt,
+    updated_at: doc.updatedAt,
+    nature: doc.nature,
+    groupement: doc.groupement,
+    title: doc.title,
+    short_title: doc.shortTitle,
+    title_slug: doc.titleSlug || slugify(doc.title),
+    short_title_slug: doc.shortTitleSlug || (doc.shortTitle ? slugify(doc.shortTitle) : null),
+    names: doc.names,
+    object: doc.object,
+    social_object1: doc.socialObject1,
+    social_object2: doc.socialObject2,
+    address_complement: doc.addressComplement,
+    address_number: doc.addressNumber,
+    address_repetition: doc.addressRepetition,
+    address_type: doc.addressType,
+    address_street: doc.addressStreet,
+    address_distribution: doc.addressDistribution,
+    address_insee_code: doc.addressInseeCode,
+    address_postal_code: doc.addressPostalCode,
+    address_department_code: doc.addressDepartmentCode,
+    address_department_name: doc.addressDepartmentName,
+    address_region: doc.addressRegion,
+    address_city: doc.addressCity,
+    management_declarant: doc.managementDeclarant,
+    management_complement: doc.managementComplement,
+    management_street: doc.managementStreet,
+    management_distribution: doc.managementDistribution,
+    management_postal_code: doc.managementPostalCode,
+    management_city: doc.managementCity,
+    management_country: doc.managementCountry,
+    director_civility: doc.directorCivility,
+    website: doc.website,
+    observation: doc.observation,
+    sync_at: doc.syncAt,
+    source: doc.source,
+  } as PgOrganization;
+  return obj;
+};
