+++ api/src/jobs/export-stats-to-pg/utils/account.ts
@@ -126,7 +126,7 @@ const handler = async () => {
         });
         scrollId = body._scroll_id;
         data = body.hits.hits as { _id: string; _source: Stats }[];
-        console.log(`[Accounts] Total hits ${body.hits.total.value}, scrollId ${scrollId}`);
+        console.log(`[Accounts] Total hits ${body.hits.total.value}`);
       }
 
       if (data.length === 0) {
