+++ api/src/jobs/export-stats-to-pg/utils/click.ts
@@ -153,7 +153,7 @@ const handler = async () => {
         });
         scrollId = body._scroll_id;
         data = body.hits.hits as { _id: string; _source: Stats }[];
-        console.log(`[Clicks] Total hits ${body.hits.total.value}, scrollId ${scrollId}`);
+        console.log(`[Clicks] Total hits ${body.hits.total.value}`);
       }
 
       if (data.length === 0) {
